@@ -1,18 +1,15 @@
 <script>
     import Offcanvas from './Offcanvas.svelte';
     import { assets } from '$app/paths';
-    import { page } from '$app/stores';
-
+    import { page } from '$app/state';
     import { PUBLIC_CACHE } from '$env/static/public';
     import { dev } from '$app/environment';
 
-    export let devMode = dev;
-
-    export let data;
+    let { data, children } = $props();
 </script>
 
 <svelte:head>
-    {#if !devMode}
+    {#if !dev}
         <script
             src="{assets}/analytics.js?{PUBLIC_CACHE}"
             onload={`analytics('G-CFB78X3F60', { anonymize_ip: true })`}
@@ -27,11 +24,11 @@
     </div>
 </div>
 
-<slot />
+{@render children?.()}
 
 <Offcanvas navigation={data.navigation} />
 
-{#if !$page.route.id?.startsWith('/docs')}
+{#if !page.route.id?.startsWith('/docs')}
     <div>
         <script>
             sessionStorage.setItem('scroll', 0);
