@@ -1,6 +1,6 @@
 <script>
     import Header from '../Header.svelte';
-    export let data;
+    let { data } = $props();
 </script>
 
 <svelte:head>
