@@ -1,7 +1,7 @@
 <script>
     import Header from '../Header.svelte';
     import { assets } from '$app/paths';
-    export let data;
+    let { data } = $props();
 </script>
 
 <svelte:head>
