@@ -1,9 +1,9 @@
 <script>
-    import { page } from '$app/stores';
+    import { page } from '$app/state';
 </script>
 
 <div class="uk-section-primary tm-section-texture">
-    <navbar class="uk-navbar-transparent" />
+    <navbar class="uk-navbar-transparent"></navbar>
 
     <div
         class="uk-section uk-section-small uk-flex uk-flex-middle uk-text-center"
@@ -11,9 +11,11 @@
     >
         <div class="uk-width-1-1">
             <div class="uk-container">
-                <h1 class="uk-heading-2xlarge uk-margin-medium uk-text-center">{$page.status}</h1>
+                <h1 class="uk-heading-2xlarge uk-margin-medium uk-text-center">
+                    {page.status}
+                </h1>
 
-                <p class="uk-text-large">{$page.error.message}</p>
+                <p class="uk-text-large">{page.error.message}</p>
             </div>
         </div>
     </div>
