@@ -1,22 +1,24 @@
 <script>
-    import { page } from '$app/stores';
+    import { page } from '$app/state';
     import { assets, base } from '$app/paths';
     import Algolia from './Algolia.svelte';
+
+    let { ...props } = $props();
 </script>
 
 <div
-    class={`uk-navbar-container tm-navbar-container ${$$props.class ?? ''}`}
-    uk-sticky={`media: 960; target-offset: 0; ${$$props['uk-sticky'] ?? ''}`}
+    class={`uk-navbar-container tm-navbar-container ${props.class ?? ''}`}
+    uk-sticky={`media: 960; target-offset: 0; ${props['uk-sticky'] ?? ''}`}
 >
     <div class="uk-container uk-container-expand">
         <nav class="uk-navbar" uk-navbar>
             <div class="uk-navbar-left">
                 <a
                     class="uk-navbar-item uk-logo"
-                    class:uk-active={$page.route.id === '/'}
+                    class:uk-active={page.route.id === '/'}
                     href="{base}/"
                     aria-label="Back to Home"
-                    aria-current={$page.route.id === '/' ? 'page' : null}
+                    aria-current={page.route.id === '/' ? 'page' : null}
                 >
                     <img
                         src="{assets}/images/uikit-logo.svg"
@@ -33,25 +35,25 @@
                     <Algolia />
                 </div>
                 <ul class="uk-navbar-nav uk-visible@m">
-                    <li class:uk-active={$page.route.id === '/pro'}>
+                    <li class:uk-active={page.route.id === '/pro'}>
                         <a
                             href="{base}/pro"
-                            aria-current={$page.route.id === '/pro' ? 'page' : null}
+                            aria-current={page.route.id === '/pro' ? 'page' : null}
                         >
                             Pro
                         </a>
                     </li>
-                    <li class:uk-active={$page.route.id?.startsWith('/docs')}>
+                    <li class:uk-active={page.route.id?.startsWith('/docs')}>
                         <a
                             href="{base}/docs/introduction"
-                            aria-current={$page.route.id?.startsWith('/docs') ? 'page' : null}
+                            aria-current={page.route.id?.startsWith('/docs') ? 'page' : null}
                             >Documentation</a
                         >
                     </li>
-                    <li class:uk-active={$page.route.id === '/changelog'}>
+                    <li class:uk-active={page.route.id === '/changelog'}>
                         <a
                             href="{base}/changelog"
-                            aria-current={$page.route.id === '/changelog' ? 'page' : null}
+                            aria-current={page.route.id === '/changelog' ? 'page' : null}
                             >Changelog</a
                         >
                     </li>
@@ -59,12 +61,12 @@
 
                 <div class="uk-navbar-item uk-visible@m">
                     <a
-                        class:uk-active={$page.route.id === '/download'}
+                        class:uk-active={page.route.id === '/download'}
                         class="uk-button uk-button-default tm-button-default uk-icon"
                         href="{base}/download"
-                        aria-current={$page.route.id === '/download' ? 'page' : null}
+                        aria-current={page.route.id === '/download' ? 'page' : null}
                     >
-                        Download <canvas uk-icon="download" width="20" height="20" />
+                        Download <canvas uk-icon="download" width="20" height="20"></canvas>
                     </a>
                 </div>
 
@@ -74,7 +76,7 @@
                     href="#offcanvas"
                     uk-toggle
                     aria-label="Open menu"
-                />
+                ></a>
             </div>
         </nav>
     </div>
